write some comments, maybe no bugs before lab4

+++ libsponge/tcp_receiver.cc
@@ -17,11 +17,14 @@ void TCPReceiver::segment_received(const TCPSegment &seg) {
             return ;
         _syn_flag = true;
         _isn = header.seqno;
+        // 这里不能直接 return，它的测试数据中 syn 为 true 时是可以带有数据的
+        //return;
     }
     string data = seg.payload().copy();
     _check_point = unwrap(header.seqno, _isn, _check_point);
-    // 这里当 header.syn 为 true 时也要占据一个 seqno ，因此有下面的判断，并且要加一
-    uint64_t index = header.syn ? _check_point : _check_point - 1;
+    // 这里当 header.syn 为 true 时也要占据一个 seqno，而 push_substring 中的 index 是 stream index 而不是 absolute seqno
+    // 因此，当 syn 为 true 时不需要减一，而当 syn 为 false 时需要减一
+    size_t index = header.syn ? _check_point : _check_point - 1;
     _reassembler.push_substring(data, index, header.fin);
 }
